Fix auth tests
Now that we have set up Develocity in the build, there will be an access
key available in the environment when executing the tests. Because
there's an access key available, the test that exercises anonymous
access started failing, because it picked up the access key from the
environment.

This commit fixes the authentication tests so that potential access keys
are removed from the environment before each test case and restored
after each test case.