Add npm-setup GitHub Action (#138)
* Initial implementation of `setup-npm`

This new GitHub Action allows setting up Develocity for npm. The action
works by:

 - installing the Develocity build agent globally
 - creating a wrapper script around `npm` which injects the required
   `NODE_OPTIONS`.
 - prepend the npm wrapper script to `PATH`.

While GitHub Actions supports setting environment variables that persist
across build steps, setting `NODE_OPTIONS` is explicitly forbidden. The
wrapper script that we inject allows us to circumvent that limitation.

* Move post-publishing to new abstract class

Previously, all the post-publishing support lived directly in the
`BuildTool` abstract class. While this made sense when only Maven was
supported, after the introduction of npm, we need a different class
because npm does not support post-publishing yet.

* Parameterize build tool tests

This allows to reuse the same tests we have for Maven with npm as well.

* Enable Jest reporter auto-injection

When instrumenting a project that wasn't configured to collect test
results with Develocity, Jest reporter auto-injection will allow
Develocity to discover that Jest is used, inject the reporter
automatically, and collect test results without user intervention.

This requires the (currently unreleased) Develocity npm agent 3.0.0 and
newer.

* Expand install location once

* Remove npm post action

It's meant to upload Build Scan data that was not uploaded yet. This is
not supported in npm and therefore should be removed.

* Remove unused inputs and outputs

These are not used or supported by the npm action.

* Support `develocity-allow-untrusted-server` in npm

* Document npm action in readme

* Run setup-npm action in CI build

* Fix auth tests

Now that we have set up Develocity in the build, there will be an access
key available in the environment when executing the tests. Because
there's an access key available, the test that exercises anonymous
access started failing, because it picked up the access key from the
environment.

This commit fixes the authentication tests so that potential access keys
are removed from the environment before each test case and restored
after each test case.

* Address review feedback

* Mention pacote version configuration option

* Fix npm action inputs table in README

* Add option to override npm agent download URL

This will allow us to install unpublished versions of the build agent
using the action.

* Create npx wrapper as well

In the previous version, only `npm` was wrapped to set the required
NODE_OPTIONS, which means that we were not capturing Build Scan data for
`npx` executions.

With this patch, a wrapper script for `npx` is created as well.

* Use npm-setup in CI

* Fix globs in npm scripts

They need to be quoted in order to be expanded by the underlying tools
being called.

* Fix linting and formatting issues

Having fixed the linter and formatter, we now see some errors that were
not surfaced before.

* Fix typo

* Split npm-specific inputs from shared inputs